requirements.txt
anthropic==0.58.2
av==15.0.0
dataclasses_json==0.6.7
datasets
decord==0.6.0
Flask==3.1.1
flask_cors==6.0.1
huggingface_hub==0.30.2
hydra-core==1.3.2
imagesize==1.4.1
iopath==0.1.9
matplotlib==3.10.3
numpy==2.3.1
omegaconf==2.3.0
openai==1.97.1
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.3.1
Pillow==11.3.0
protobuf==6.31.1
pycocotools==2.0.8
PyYAML==6.0.2
PyYAML==6.0.2
ray==2.43.0
Requests==2.32.4
rtpt==0.0.4
setuptools==75.8.0
scikit-image
submitit==1.5.3
tensorboardX==2.6.2.2
tensordict==0.9.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
vose==0.0.2
einops
timm
accelerate
qwen_vl_utils
python-Levenshtein